+++ RELEASE_NOTES.md
@@ -43,12 +43,14 @@ This version adds new locales and makes the library build on more platforms.
 * `KeyboardAction` `isSpace` has been removed.
 * `KeyboardAction` `standardTextDocumentProxyInputAction` has been merged with `standardTextDocumentProxyAction`.
 * `KeyboardAppearance` `systemKeyboardButtonStyle(for:)` has been renamed to `buttonStyle(for:)`.
+* `KeyboardFeedbackHandler` no longer has any action provider-based logic and no typealiases.
 * `KeyboardInputViewController` `needsInputModeSwitchKeyOverride` has been removed.
 * `StandardKeyboardActionHandler` `GestureAction` has been removed.
+* `StandardKeyboardFeedbackHandler` `triggerFeedbackForLongPressOnSpaceDragGesture` has been removed.
+* `StandardKeyboardFeedbackHandler` `shouldTriggerFeedback` has been moved to `StandardKeyboardActionHandler`.
 * `SystemKeyboardButtonStyle` has been renamed to `KeyboardButtonStyle`.
 * `SystemKeyboardButtonBorderStyle` has been renamed to `KeyboardButtonBorderStyle`.
 * `SystemKeyboardButtonShadowStyle` has been renamed to `KeyboardButtonShadowStyle`.
-* `StandardKeyboardFeedbackHandler` `triggerFeedbackForLongPressOnSpaceDragGesture` has been removed.